Added advanced options and general cleanup
This commit adds advanced options to the Mixin. Notably, the end user
can now opt-in to enabling 'scroll spy', which is the constant watching
of the Component to determine whether or not it has entered or left the
viewport. If the user opts-out (which is a default), the Mixin will only
watch the Component up to the point it enters the viewport, at which it
will unbind its event listeners registered on the window and document,
and remove its observers.