Register the Component's elementId along with it's rAF call
This commit enables `viewportSpy: false` to work when rAF is enabled.
The Mixin essentially stores an object with elementIds as keys for all Objects
with the Mixin, and with their respective requestAnimationFrame ids as
values. This allows for `#_unbindListeners` to cancel the appropriate
animation frame, based on that Component's elementId.