Prevent unsetting viewportEntered when not spying
When scrolling through a large list of components with the InViewport
mixin, the `_triggerDidAccessViewport` method can set the `viewportEntered`
property to false after listeners for the component have been unbound, causing the
components to remain stuck in their outside viewport state.
`viewportEntered` should only be set back to false if the spy is being used.