bug: handle subpixel values in viewport dims
Rounds the dimensions on the DOMRect of the in-viewport element since
`jQuery.innerWidth()` (or `height`) return integers.
`Element.getBoundingClientRect()` can return subpixels (as floating
point numbers) that if not handled will result in elements not being
considered in the viewport by the `is-in-viewport` helper (even though
they are).

This is particularly important when the user has zoomed the window,
in which case the browser may scale the size of the viewport, resulting
in the dimensions having fractional parts, e.g., the width might be
`1451.1112060546875`.

Closes #96