Run service teardown code in willDestroy()
Put the service teardown code in willDestroy() instead of destroy() so it runs at the expected time. destroy() schedules a call to willDestroy() on the run loop, so they happen at different times, and willDestroy() is the hook where teardown code is expected to live. Starting in Ember 3.20 having this code in destroy() can cause test teardown bugs if a modifier calls into the service to stop watching, because things happen in this order:

1. Service's destroy() method is called (by the container)
2. Modifiers' willRemove() hooks are called
3. Service's willDestroy() hook is called

So the service's teardown code needs to live in willDestroy() or else it will have already torn down by step (2) and calls into it from component/modifier teardown code can cause errors.